requirements.txt
python-telegram-bot
google-generativeai
openai
anthropic
pyyaml
rich
requests
questionary
googlesearch-python
trafilatura
